Refactor(wonji/salelists): 변수명 변경, 함수명 변경, 판매목록 조회시 실패하게 되면 msg띄우도록 구현
