+++ app/app.js
@@ -18,7 +18,7 @@ app.use(cookieParser());
 
 const view = require("./src/apis/view");
 const root = require("./src/apis/root");
-const sale = require("./src/apis/saleList/routes");
+const sale = require("./src/apis/sale-list/routes");
 
 app.use("/", view);
 app.use("/api/", root);
