+++ app/src/apis/sale-list/routes.js
@@ -2,9 +2,9 @@
 
 const express = require("express");
 const router = express.Router();
-const ctrl = require("./saleList.ctrl");
+const ctrl = require("./sale-list.ctrl");
 
-router.get("/", ctrl.output.findSaleList);
+router.get("/", ctrl.output.read);
 router.post("/", ctrl.process.update);
 
 module.exports = router;
